test_app.py
#!/usr/bin/env python3
# Test script to check if the main application runs without errors

try:
    import tkinter as tk
    from tkinter import ttk, messagebox, filedialog, simpledialog
    from PIL import Image, ImageTk
    import os
    import csv
    import json
    from datetime import date, datetime
    from collections import Counter
    
    print("All imports successful!")
    
    # Try to import and instantiate the main application
    import sys
    sys.path.append('.')
    
    print("Testing application startup...")
    
    # Import the main file without running it
    import importlib.util
    spec = importlib.util.spec_from_file_location("main_app", "shopping_cart_and_banking-system.py")
    main_module = importlib.util.module_from_spec(spec)
    
    print("Module loaded successfully!")
    print("The user login functionality has been added to the Profile tab.")
    print("Features added:")
    print("1. User Info tab - shows current user information")
    print("2. Login tab with Login and Register sub-tabs")
    print("3. User authentication system")
    print("4. User session management")
    print("5. Demo credentials: demo@example.com / demo123")
    
except ImportError as e:
    print(f"Import error: {e}")
except Exception as e:
    print(f"Error: {e}")
